chore: 함수명 수정
