+++ android/core/ui/src/main/kotlin/com/andlife/ui/component/guestbook/GuestBookItem.kt
@@ -1,7 +1,5 @@
 package com.andlife.ui.component.guestbook
 
-import android.view.ViewGroup
-import androidx.annotation.OptIn
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.snap
@@ -34,32 +32,22 @@ import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.boundsInWindow
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.media3.common.Player
-import androidx.media3.common.util.UnstableApi
 import coil3.compose.AsyncImage
 import coil3.compose.SubcomposeAsyncImage
 import com.andlife.designsystem.preview.PreviewTheme
@@ -68,7 +56,6 @@ import com.andlife.designsystem.theme.NachoSpacing
 import com.andlife.designsystem.theme.NachoStroke
 import com.andlife.designsystem.theme.NachoTheme
 import com.andlife.media.audio.AudioPlaybackState
-import com.andlife.media.video.AutoVideoPlayer
 import com.andlife.media.video.AutoVideoPlayerPool
 import com.andlife.media.video.FakeAutoVideoPlayerPool
 import com.andlife.model.common.AuthorUiModel
@@ -77,14 +64,11 @@ import com.andlife.model.guestbook.GuestBookMediaUiModel
 import com.andlife.model.guestbook.GuestBookUiModel
 import com.andlife.model.guestbook.MediaUiType
 import com.andlife.ui.R
-import com.andlife.ui.component.icon.PlayerThumbnailIcon
 import com.andlife.ui.component.media.MediaOverlay
-import com.andlife.ui.component.media.video.VideoPlaybackControlOverlay
 import com.andlife.ui.util.toFormatDuration
 import com.andlife.ui.util.toRelativeTimeString
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.delay
 import kotlinx.datetime.LocalDateTime
 import com.andlife.designsystem.R as designR
 
@@ -427,7 +411,7 @@ private fun GuestBookItemVisualMediaSection(
             ) {
                 when (media.type) {
                     MediaUiType.VIDEO -> {
-                        VideoPlayerContainer(
+                        GuestBookVideoPlayer(
                             guestBookId = guestBookId,
                             videoUrl = media.url,
                             thumbnailUrl = media.thumbnailUrl,
@@ -486,255 +470,6 @@ private fun GuestBookItemVisualMediaSection(
                 shape = NachoTheme.shapes.medium,
             )
         }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun VideoPlayerContainer(
-    guestBookId: Long,
-    videoUrl: String,
-    thumbnailUrl: String?,
-    totalDurationSeconds: Int?,
-    shouldPlay: Boolean,
-    isFullscreen: Boolean,
-    videoPlayerPool: AutoVideoPlayerPool,
-    onFullscreenClick: (String, String?, Rect) -> Unit,
-    onPlayVideoClick: (String) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var isVideoReady by remember(videoUrl) { mutableStateOf(false) }
-    var currentPositionMs by remember(videoUrl) { mutableLongStateOf(0L) }
-    var totalDurationMs by remember(videoUrl) { mutableLongStateOf((totalDurationSeconds?.times(1000L)) ?: 0L) }
-    var isControlVisible by remember(videoUrl) { mutableStateOf(false) }
-    var isSeeking by remember(videoUrl) { mutableStateOf(false) }
-    var seekPositionMs by remember(videoUrl) { mutableLongStateOf(0L) }
-    val isMuted by videoPlayerPool.isMuted.collectAsStateWithLifecycle()
-    var containerBounds by remember { mutableStateOf(Rect.Zero) }
-
-    val displayPositionMs = if (isSeeking) seekPositionMs else currentPositionMs
-    val progress = if (totalDurationMs > 0) {
-        (displayPositionMs.toFloat() / totalDurationMs.toFloat()).coerceIn(0f, 1f)
-    } else {
-        0f
-    }
-    val timeText = "${(displayPositionMs / 1000).toInt().toFormatDuration()} / ${
-        (totalDurationMs / 1000).toInt().toFormatDuration()
-    }"
-
-    LaunchedEffect(shouldPlay, videoUrl) {
-        if (shouldPlay) {
-            videoPlayerPool.playPlayer(videoUrl, guestBookId)
-        } else {
-            videoPlayerPool.pausePlayer(videoUrl)
-        }
-    }
-
-    LaunchedEffect(isControlVisible) {
-        if (isControlVisible) {
-            delay(3000L)
-            isControlVisible = false
-        }
-    }
-
-    LaunchedEffect(shouldPlay) {
-        if (!shouldPlay) isControlVisible = false
-    }
-
-    if (shouldPlay) {
-        val currentPlayer = remember(videoUrl) { videoPlayerPool.getPlayer(videoUrl) }
-
-        DisposableEffect(currentPlayer, videoUrl) {
-            val listener = object : Player.Listener {
-                override fun onRenderedFirstFrame() {
-                    isVideoReady = true
-                }
-
-                override fun onPlaybackStateChanged(state: Int) {
-                    if (state == Player.STATE_READY && currentPlayer.exoPlayer.playWhenReady) {
-                        isVideoReady = true
-                        totalDurationMs = currentPlayer.exoPlayer.duration.coerceAtLeast(0L)
-                    }
-                }
-            }
-            currentPlayer.exoPlayer.addListener(listener)
-            if (currentPlayer.exoPlayer.playbackState == Player.STATE_READY) {
-                isVideoReady = true
-                totalDurationMs = currentPlayer.exoPlayer.duration.coerceAtLeast(0L)
-            }
-            onDispose {
-                currentPlayer.exoPlayer.removeListener(listener)
-            }
-        }
-
-        LaunchedEffect(isVideoReady) {
-            if (!isVideoReady) {
-                currentPositionMs = 0L
-                return@LaunchedEffect
-            }
-            while (true) {
-                if (!isSeeking) {
-                    currentPositionMs = currentPlayer.exoPlayer.currentPosition.coerceAtLeast(0L)
-                }
-                delay(250L)
-            }
-        }
-
-        VideoPlayerContent(
-            currentPlayer = currentPlayer,
-            thumbnailUrl = thumbnailUrl,
-            progress = progress,
-            timeText = timeText,
-            isControlVisible = isControlVisible,
-            isMuted = isMuted,
-            isVideoReady = isVideoReady,
-            isFullscreen = isFullscreen,
-            onVideoClick = { isControlVisible = !isControlVisible },
-            onPlayVideoClick = { onPlayVideoClick(videoUrl) },
-            onContainerPositioned = { containerBounds = it },
-            onSeekValueChange = { newValue ->
-                if (!isSeeking) currentPlayer.pause()
-                isSeeking = true
-                seekPositionMs = (newValue * totalDurationMs).toLong()
-            },
-            onSeekValueChangeFinished = {
-                currentPlayer.seekTo(seekPositionMs)
-                currentPlayer.play()
-                isSeeking = false
-            },
-            onMuteToggle = { videoPlayerPool.toggleMute() },
-            onFullscreenClick = { onFullscreenClick(videoUrl, thumbnailUrl, containerBounds) }
-        )
-    } else {
-        VideoPlayerContent(
-            currentPlayer = null,
-            thumbnailUrl = thumbnailUrl,
-            progress = 0f,
-            timeText = "",
-            isControlVisible = false,
-            isMuted = isMuted,
-            isVideoReady = false,
-            isFullscreen = false,
-            onVideoClick = {},
-            onPlayVideoClick = { onPlayVideoClick(videoUrl) },
-            onContainerPositioned = {},
-            onSeekValueChange = {},
-            onSeekValueChangeFinished = {},
-            onMuteToggle = { videoPlayerPool.toggleMute() },
-            onFullscreenClick = {}
-        )
-    }
-}
-
-@Composable
-private fun VideoPlayerContent(
-    currentPlayer: AutoVideoPlayer?,
-    thumbnailUrl: String?,
-    progress: Float,
-    timeText: String,
-    isControlVisible: Boolean,
-    isMuted: Boolean,
-    isVideoReady: Boolean,
-    isFullscreen: Boolean,
-    onVideoClick: () -> Unit,
-    onPlayVideoClick: () -> Unit,
-    onContainerPositioned: (Rect) -> Unit,
-    onSeekValueChange: (Float) -> Unit,
-    onSeekValueChangeFinished: () -> Unit,
-    onMuteToggle: () -> Unit,
-    onFullscreenClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val thumbnailAlpha by animateFloatAsState(
-        targetValue = if (isVideoReady) 0f else 1f,
-        animationSpec = tween(durationMillis = 200),
-    )
-
-    Box(
-        modifier = modifier
-            .fillMaxSize()
-            .background(Color.Black)
-            .onGloballyPositioned { coordinates ->
-                onContainerPositioned(coordinates.boundsInWindow())
-            }
-            .then(
-                if (isVideoReady) Modifier.clickable { onVideoClick() }
-                else Modifier
-            )
-    ) {
-        currentPlayer?.let {
-            if (!isFullscreen) {
-                VideoSurface(
-                    autoPlayer = it,
-                    modifier = Modifier.fillMaxSize(),
-                )
-            } else {
-                if (thumbnailUrl != null) {
-                    AsyncImage(
-                        model = thumbnailUrl,
-                        contentDescription = null,
-                        modifier = Modifier.fillMaxSize(),
-                        contentScale = ContentScale.Fit,
-                    )
-                }
-            }
-        }
-
-        if (thumbnailUrl != null && thumbnailAlpha > 0f) {
-            VideoThumbnail(
-                thumbnailUrl = thumbnailUrl,
-                onPlayVideoClick = onPlayVideoClick,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .alpha(thumbnailAlpha),
-            )
-        }
-
-        if (currentPlayer != null) {
-            VideoPlaybackControlOverlay(
-                progress = progress,
-                timeText = timeText,
-                isControlVisible = isControlVisible,
-                isMuted = isMuted,
-                onSeekValueChange = onSeekValueChange,
-                onSeekValueChangeFinished = onSeekValueChangeFinished,
-                onMuteToggle = onMuteToggle,
-                onFullscreenClick = onFullscreenClick,
-            )
-        }
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-fun VideoSurface(
-    autoPlayer: AutoVideoPlayer,
-    modifier: Modifier = Modifier,
-) {
-    AndroidView(
-        factory = {
-            autoPlayer.playerView.apply {
-                (parent as? ViewGroup)?.removeView(this)
-            }
-        },
-        modifier = modifier,
-    )
-}
-
-@Composable
-private fun VideoThumbnail(
-    thumbnailUrl: String?,
-    onPlayVideoClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(modifier = modifier.clickable { onPlayVideoClick() }) {
-        AsyncImage(
-            model = thumbnailUrl,
-            contentDescription = stringResource(R.string.desc_video_thumbnail),
-            modifier = Modifier.fillMaxSize(),
-            contentScale = ContentScale.Fit,
-        )
-        PlayerThumbnailIcon(modifier = Modifier.align(Alignment.Center))
     }
 }
